# api/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "FieldViolation": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "code": {
            "type": "string"
          },
          "currentRevision": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "expectedRevision": {
            "type": "string"
          },
          "fieldViolations": {
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            },
            "type": "array"
          },
          "instance": {
            "type": "string"
          },
          "remediation": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "status": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "type": "object"
      },
...
      "Session": {
        "properties": {
          "actor": {
            "$ref": "#/components/schemas/SessionActor"
          },
          "csrf": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "SessionActor": {
        "properties": {
          "class": {
            "type": "string"
          },
          "groups": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "token",
        "description": "Required for non-loopback peers. Loopback (127.0.0.1 / ::1) may omit the token (Q-AUTH / dev-loopback-unauth).",
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "description": "Browser session cookie. CSRF header X-LabDNS-CSRF is required on cookie-authenticated non-GET requests.",
        "in": "cookie",
        "name": "labdns_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "REST adapter for the shared LabDNS capability registry. Generated from internal/capabilities and model.Spec. Do not edit by hand.",
    "title": "LabDNS Management API",
    "version": "v1",
    "x-generated-by": "internal/control/rest.RenderOpenAPI; DO NOT EDIT."
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/session": {
      "delete": {
        "description": "Browser session cookie and CSRF secret. REST only.",
        "operationId": "delete_v1_session",
        "parameters": [
          {
            "description": "required when authenticating with cookie labdns_session; ignored for Authorization: Bearer.",
            "in": "header",
            "name": "X-LabDNS-CSRF",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session create/get/delete"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "application/problem+json domain error"
          }
        },
        "summary": "Session create/get/delete",
        "tags": [
          "session"
        ],
        "x-capability-id": "session",
        "x-idempotent": true,
        "x-mutating": false
      },
      "get": {
        "description": "Browser session cookie and CSRF secret. REST only.",
        "operationId": "get_v1_session",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "description": "Session create/get/delete"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "application/problem+json domain error"
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "Session create/get/delete",
        "tags": [
          "session"
        ],
        "x-capability-id": "session",
        "x-idempotent": true,
        "x-mutating": false
      },
      "post": {
        "description": "Browser session cookie and CSRF secret. REST only.",
        "operationId": "post_v1_session",
        "parameters": [
          {
            "description": "required when authenticating with cookie labdns_session; ignored for Authorization: Bearer.",
            "in": "header",
            "name": "X-LabDNS-CSRF",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "description": "Session create/get/delete"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "application/problem+json domain error"
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "Session create/get/delete",
        "tags": [
          "session"
        ],
        "x-capability-id": "session",
        "x-idempotent": true,
        "x-mutating": false
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "servers": [
    {
      "description": "Management listener (default address :8080)",
      "url": "/"
    }
  ],
...
}